This is hopefully the final commit
In this commit I have created another recipe page.
The format is same as previous and I have
included the hyperlink in the index file.

The image is stored in the images folder.

I have also cropped the jhalmuri image in the folder.

A summary of the project is upfdated in the
README file.